move client and parser modules into a slack2discord package
and update imports accordingly
and this allows us to simplify some naming a bit
