+++ slack2discord.py
@@ -4,18 +4,14 @@
 # Enough chaanges led to a hard fork, it is now slack2discord, by Rich Fromm
 
 import asyncio
-import discord
-from discord.ext import commands
 import logging
 from sys import argv, exit
 
-#from slack2discord.client import DiscordClient
-#from .client import DiscordClient
-#from . import DiscordClient
-import client
-#from slack2discord.parser import SlackParser
-#from .parser import SlackParser
-import parser
+import discord
+from discord.ext import commands
+
+from slack2discord.client import DiscordClient
+from slack2discord.parser import SlackParser
 
 
 logger = logging.getLogger('slack2discord')
@@ -26,9 +22,9 @@ def post_to_discord(token, channel_id, parsed_messages, verbose):
     Iterate through the results of previously parsing the JSON file from a Slack export and post
     each message to Discord in the channel corresponding to the given id. Threading is preserved.
     """
-    discord_client = client.DiscordClient(channel_id, parsed_messages, verbose)
+    client = DiscordClient(channel_id, parsed_messages, verbose)
     # if Ctrl-C is pressed, we do *not* get a KeyboardInterrupt b/c it is caught by the run() loop in the discord client
-    discord_client.run(token)
+    client.run(token)
 
 if __name__ == '__main__':
     # Normally logging gets set up automatically when discord.Client.run() is called.
@@ -47,10 +43,10 @@ if __name__ == '__main__':
     # XXX this should be an arg, for now just edit here
     verbose = False
 
-    slack_parser = parser.SlackParser(verbose)
-    slack_parser.parse_json_slack_export(filename)
+    parser = SlackParser(verbose)
+    parser.parse_json_slack_export(filename)
 
-    post_to_discord(token, channel_id, slack_parser.parsed_messages, verbose)
+    post_to_discord(token, channel_id, parser.parsed_messages, verbose)
     # XXX return values of asyncio functions are tricky, don't worry about it for now
     logger.info("Discord import complete (may or may not have been successful)")
     exit(0)
