+++ slack2discord/client.py
