+++ slack2discord/parser.py
